refactor: move toggle in another directory
